requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
cssselect2==0.9.0
Django==5.2.11
django-admin-sortable2==2.3.1
django-environ==0.13.0
django-filer==3.4.4
django-polymorphic==4.11.1
easy-thumbnails==2.10.1
lxml==6.0.2
mysqlclient==2.2.8
pillow==12.1.1
reportlab==4.4.10
sqlparse==0.5.5
svglib==1.5.1
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.3
webencodings==0.5.1
